removed favoriteBeer, interestedInPronoun, interestedInRelation, renamed beer seeds file, reworked seed.js, renamed relationshipTypes, added to userData, added to pronounData, added to relationshipTypeData
