+++ .env
@@ -1,3 +1,4 @@
 DB_NAME=mixNmatch_db
-DB_PASSWORD=newpassword
 DB_USER=root
+DB_PASSWORD=newpassword
+
