+++ Main/config/connection.js
@@ -18,3 +18,21 @@ if(process.env.JAWSDB_URL){
 }
 
 module.exports = sequelize;
+
+
+// Daniel used the following to test today
+// const Sequelize = require('sequelize');
+// require('dotenv').config();
+
+// const sequelize = new Sequelize(
+//   process.env.DB_NAME,
+//   process.env.DB_USER,
+//   process.env.DB_PASSWORD,
+//   {
+//     host: 'localhost',
+//     dialect: 'mysql',
+//     port: 3306,
+//   }
+// );
+
+// module.exports = sequelize;
