+++ Main/models/index.js
@@ -1,59 +1,33 @@
 // const Conversations = require('../models/conversations');
 // const Matches = require('./matches');
-const InterestedInPronoun = require('./interestedInPronoun');
-const Pronoun = require('./pronoun');
-const InterestedInRelation = require('./interestedInRelation');
-const RelationshipType = require('./relationshipType');
+const Pronoun = require('./pronouns');
+const RelationshipType = require('./relationshipTypes');
 const User = require('./user');
 const Beers = require('./beers');
-const FavoriteBeer = require('./favoriteBeer');
 const Photo = require('./photo');
 
-Pronoun.belongsTo(User, {
-  foreignKey: 'user_id',
-});
 
-InterestedInPronoun.belongsTo(User, {
-  foreignKey: 'user_id',
+Pronoun.belongsTo(User, {
+  foreignKey: 'pronouns_name',
 });
 
 RelationshipType.belongsTo(User, {
-  through: {
-    model: InterestedInRelation,
-    unique: false,
-  },
-
-  as: 'relationshipType_id',
+  foreignKey: 'relationshipType_name',
 });
 
-InterestedInRelation.belongsTo(User, {
-  foreignKey: 'relationshipType_id',
-});
-
-Beers.belongsToMany(User, {
-  through: {
-    model: FavoriteBeer,
-    unique: false,
-  },
-
-  as: 'beers_id',
-});
-
-FavoriteBeer.belongsTo(User, {
-  foreignKey: 'beers_id',
+Beers.belongsTo(User, {
+ foreignKey: 'beers_name'
 });
 
 // User.hasMany(Matches, {
-
 // });
 
 // User.hasMany(Conversations, {
-
 // });
 
-Photo.belongsTo(User, {
-  foreignKey: 'user_id',
-});
+// Photo.belongsTo(User, {
+//   foreignKey: 'user_id',
+// });
 
 User.belongsToMany(User, {
   through: {
@@ -67,11 +41,8 @@ User.belongsToMany(User, {
 module.exports = {
   User,
   Pronoun,
-  InterestedInPronoun,
-  InterestedInRelation,
   RelationshipType,
   Beers,
-  FavoriteBeer,
   // Matches,
   // Conversations,
   Photo,
