--- Main/models/interestedInRelation.js
@@ -1,38 +1,0 @@
-const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config/connection');
-// const InterestedInPronoun = require('./interestedInPronoun');
-
-// Create a new Sequelize model for books
-class interestedInRelation extends Model {}
-
-interestedInRelation.init(
-  {
-    id: {
-        type: DataTypes.INTEGER,
-        autoIncrement:true,
-        primaryKey:true,
-      },
-      user_id: {
-          type: DataTypes.INTEGER,
-          references: {
-            model: 'user',
-            key: 'id',
-          },
-      },
-      relationshipType_id: {
-          type: DataTypes.INTEGER,
-          references: {
-              model: 'relationshipType',
-              key: 'id'
-          }
-      }
-  },
-  {
-    sequelize,
-    timestamps: false,
-    underscored: true,
-    modelName: 'interestedInRelation'
-  }
-);
-
-module.exports = interestedInRelation;
