--- Main/seeds/beer.json
@@ -1,18 +1,0 @@
-[
-    {
-        "name": "Irish Dry Stout"
-    },
-    {
-        "name": "IPA"
-    },
-    {
-        "name": "Pale Lager"
-    },
-    {
-        "name": "Lager"
-    },
-    {
-        "name": "American Lager"
-    }
-
-]
