+++ Main/seeds/pronounData.json
@@ -1,3 +1,11 @@
-{
-    
-}
+[
+  {
+    "name": "He/Him"
+  },
+  {
+    "name": "She/Her"
+  },
+  {
+    "name": "They/Them"
+  }
+]
