+++ Main/seeds/relationshipTypeData.json
@@ -1,3 +1,8 @@
-{
-    
-}
+[
+  {
+    "name": "Dating"
+  },
+  {
+    "name": "Friendship"
+  }
+]
