+++ Main/seeds/seed.js
@@ -1,21 +1,34 @@
 const sequelize = require('../config/connection');
-const { User, 
-    RelationshipType,
-    Beers,
-    // Matches,
-    // Conversations,
-    Photo } = require('../models');
+const {
+  User,
+  RelationshipType,
+  Beers,
+  // Matches,
+  // Conversations,
+  Pronoun,
+  Photo
+} = require('../models');
 
 const userData = require('./userData.json');
-const beersData = require('./beerData.json');
+const beersData = require('./beersData.json');
 const photoData = require('./photoData.json');
-// const pronounData = require('./pronounData.json');
+const pronounData = require('./pronounData.json');
 const relationshipTypeData = require('./relationshipTypeData.json');
 
-const seedDatabase = async () => {
+const seedAll = async () => {
   await sequelize.sync({ force: true });
 
   await User.bulkCreate(userData, {
+    individualHooks: true,
+    returning: true,
+  });
+
+  await Pronoun.bulkCreate(pronounData, {
+    individualHooks: true,
+    returning: true,
+  });
+
+  await RelationshipType.bulkCreate(relationshipTypeData, {
     individualHooks: true,
     returning: true,
   });
@@ -30,18 +43,9 @@ const seedDatabase = async () => {
     returning: true,
   });
 
-  // await Pronoun.bulkCreate(pronounData, {
-  //   individualHooks: true,
-  //   returning: true,
-  // });
-
-  await RelationshipType.bulkCreate(relationshipTypeData, {
-    individualHooks: true,
-    returning: true,
-  });
-
+  
 
   process.exit(0);
 };
 
-seedDatabase();
+seedAll();
